% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_CellTalkTarget_cci}
\alias{get_CellTalkTarget_cci}
\title{Get miRNA-target interactions}
\usage{
get_CellTalkTarget_cci(object, simple = TRUE)
}
\arguments{
\item{object}{CellTalkTarget object after \code{\link{find_CellTalkTarget}}}

\item{simple}{Whether to show the simple results. Default is \code{TRUE}}
}
\value{
A data.frame containing all miRNA-target interactions.
}
\description{
Get simple results of miRNA-target interactions and specificity.
}
