% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_CellTalkTarget_circulating_score}
\alias{get_CellTalkTarget_circulating_score}
\title{Get circulating score of inferred miRNAs}
\usage{
get_CellTalkTarget_circulating_score(object)
}
\arguments{
\item{object}{CellTalkTarget object after \code{\link{find_CellTalkTarget}}}
}
\value{
A data.frame containing all potential circulating miRNAs.
}
\description{
Get circulating score of inferred miRNAs.
}
