% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show,CellTalkTarget-method}
\alias{show,CellTalkTarget-method}
\title{Show CellTalkTarget object}
\usage{
\S4method{show}{CellTalkTarget}(object)
}
\arguments{
\item{object}{CellTalkTarget object}
}
\value{
CellTalkTarget object
}
\description{
Show CellTalkTarget object
}
